The websocket connections seem to be closer to working
