+++ src/feature_pipeline/client.py
@@ -1,0 +1,14 @@
+import asyncio
+import websockets 
+from loguru import logger
+
+
+async def client():
+    async with websockets.connect("ws://localhost:8515") as websocket:
+        message = await websocket.recv()
+        logger.success(f"Received: {message}")
+
+
+if __name__ == "__main__":
+    asyncio.run(client())
+
